item.stories draft first commit
converted 70%
