+++ package.json
@@ -149,8 +149,8 @@
     "@storybook/react-webpack5": "^7.3.2",
     "@storybook/testing-library": "^0.2.0",
     "concurrently": "3.6.1",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.7",
     "expect": "^23.6.0",
     "nodemon": "1.18.9",
     "storybook": "^7.3.2"
